Separate from the internals class'
